server/utils/requirements.txt
scikit-learn
pymongo
python-dotenv